# openapi/api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Minimarket Plus API",
    "description": "API REST del sistema backend Minimarket Plus. Gestiona productos, categorías, carritos, inventario, ventas y usuarios con roles (cliente, cajero, administrador). Todos los endpoints de negocio requieren autenticación HTTP Basic; use el botón Authorize (credenciales de demostración en el README).",
    "contact": {
      "name": "Equipo Minimarket Plus - PBY2202",
      "email": "karla.santigut@gmail.com"
    },
    "license": {
      "name": "Uso académico - Duoc UC"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Entorno local de desarrollo"
    }
  ],
  "security": [
    {
      "basicAuth": []
    }
  ],
...
  "paths": {
...
    "/api/inventario/{id}": {
      "get": {
        "tags": [
          "Inventario"
        ],
        "summary": "Obtener un movimiento por ID",
        "description": "Busca un movimiento de inventario específico por su identificador único.",
        "operationId": "obtenerMovimientoPorId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador único del movimiento",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Movimiento encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Inventario"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado"
          },
          "404": {
            "description": "No existe un movimiento con el ID indicado"
          }
        }
      },
      "put": {
        "tags": [
          "Inventario"
        ],
        "summary": "Actualizar un movimiento de inventario",
        "description": "Modifica los datos de un movimiento existente.",
        "operationId": "actualizarMovimiento",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador único del movimiento a actualizar",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "requestBody": {
          "description": "Nuevos datos del movimiento",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Inventario"
              },
              "examples": {
                "movimientoActualizado": {
                  "description": "movimientoActualizado",
                  "value": {
                    "producto": {
                      "id": 1
                    },
                    "cantidad": 10,
                    "tipoMovimiento": "Entrada",
                    "fechaMovimiento": "2026-07-05T10:00:00.000+00:00"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "401": {
            "description": "No autenticado"
          },
          "404": {
            "description": "No existe un movimiento con el ID indicado"
          },
          "200": {
            "description": "Movimiento actualizado correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Inventario"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Inventario"
        ],
        "summary": "Eliminar un movimiento de inventario",
        "description": "Elimina el movimiento identificado por el ID.",
        "operationId": "eliminarMovimiento",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador único del movimiento a eliminar",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "401": {
            "description": "No autenticado"
          },
          "404": {
            "description": "No existe un movimiento con el ID indicado"
          },
          "204": {
            "description": "Movimiento eliminado correctamente"
          }
        }
      }
    },
...
    "/api/inventario": {
      "get": {
        "tags": [
          "Inventario"
        ],
        "summary": "Listar todos los movimientos de inventario",
        "description": "Retorna el historial completo de entradas y salidas de productos.",
        "operationId": "listarMovimientosDeInventario",
        "responses": {
          "401": {
            "description": "No autenticado"
          },
          "200": {
            "description": "Listado obtenido correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Inventario"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Inventario"
        ],
        "summary": "Registrar un movimiento de inventario",
        "description": "Registra una entrada o salida de unidades para un producto existente.",
        "operationId": "registrarMovimiento",
        "requestBody": {
          "description": "Movimiento a registrar (el producto debe existir)",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Inventario"
              },
              "examples": {
                "nuevoMovimiento": {
                  "description": "nuevoMovimiento",
                  "value": {
                    "producto": {
                      "id": 1
                    },
                    "cantidad": 10,
                    "tipoMovimiento": "Entrada",
                    "fechaMovimiento": "2026-07-05T10:00:00.000+00:00"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "401": {
            "description": "No autenticado"
          },
          "201": {
            "description": "Movimiento registrado correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Inventario"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/ventas/{id}": {
      "get": {
        "tags": [
          "Ventas"
        ],
        "summary": "Obtener una venta por ID",
        "description": "Busca una venta específica por su identificador único.",
        "operationId": "obtenerVentaPorId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador único de la venta",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "401": {
            "description": "No autenticado"
          },
          "404": {
            "description": "No existe una venta con el ID indicado"
          },
          "200": {
            "description": "Venta encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Venta"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Rol": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Identificador único del rol",
            "format": "int64",
            "readOnly": true,
            "example": 1
          },
          "nombre": {
            "type": "string",
            "description": "Nombre del rol",
            "example": "ROLE_ADMIN"
          }
        },
        "description": "Rol de autorización dentro del sistema"
      },
      "Usuario": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Identificador único del usuario",
            "format": "int64",
            "readOnly": true,
            "example": 1
          },
          "username": {
            "type": "string",
            "description": "Nombre de usuario único para autenticación",
            "example": "admin"
          },
          "password": {
            "type": "string",
            "description": "Contraseña del usuario (solo escritura, no se retorna en respuestas)",
            "writeOnly": true,
            "example": "admin123"
          },
          "roles": {
            "uniqueItems": true,
            "type": "array",
            "description": "Roles asignados al usuario",
            "items": {
              "$ref": "#/components/schemas/Rol"
            }
          }
        },
        "description": "Usuario del sistema con sus roles asociados (cliente, cajero, administrador)"
      },
      "Categoria": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Identificador único de la categoría",
            "format": "int64",
            "readOnly": true,
            "example": 1
          },
          "nombre": {
            "type": "string",
            "description": "Nombre único de la categoría",
            "example": "Bebidas"
          }
        },
        "description": "Categoría para clasificar los productos del minimarket"
      },
      "Producto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Identificador único del producto",
            "format": "int64",
            "readOnly": true,
            "example": 1
          },
          "nombre": {
            "type": "string",
            "description": "Nombre comercial del producto",
            "example": "Coca-Cola 1.5L"
          },
          "precio": {
            "type": "number",
            "description": "Precio unitario en pesos chilenos",
            "format": "double",
            "example": 1890
          },
          "stock": {
            "type": "integer",
            "description": "Unidades disponibles en bodega",
            "format": "int32",
            "example": 24
          },
          "categoria": {
            "$ref": "#/components/schemas/Categoria"
          }
        },
        "description": "Producto del catálogo del minimarket"
      },
      "Inventario": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Identificador único del movimiento",
            "format": "int64",
            "readOnly": true,
            "example": 1
          },
          "producto": {
            "$ref": "#/components/schemas/Producto"
          },
          "cantidad": {
            "type": "integer",
            "description": "Cantidad de unidades del movimiento",
            "format": "int32",
            "example": 10
          },
          "tipoMovimiento": {
            "type": "string",
            "description": "Tipo de movimiento de inventario",
            "example": "Entrada",
            "enum": [
              "Entrada",
              "Salida"
            ]
          },
          "fechaMovimiento": {
            "type": "string",
            "description": "Fecha y hora del movimiento",
            "format": "date-time",
            "example": "2026-07-05T10:00:00Z"
          }
        },
        "description": "Movimiento de inventario (entrada o salida) de un producto"
      },
...
      "Venta": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Identificador único de la venta",
            "format": "int64",
            "readOnly": true,
            "example": 1
          },
          "usuario": {
            "$ref": "#/components/schemas/Usuario"
          },
          "fecha": {
            "type": "string",
            "description": "Fecha y hora de la venta",
            "format": "date-time",
            "example": "2026-07-05T14:30:00Z"
          }
        },
        "description": "Venta realizada en el minimarket, con su usuario, fecha y detalles"
      },
...
    },
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "description": "Autenticación HTTP Basic contra los usuarios registrados en H2",
        "name": "basicAuth",
        "scheme": "basic"
      }
    }
  }
}